Rename typo 'timer_contdown_widget' to 'timer_countdown_widget'
